# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VPN Manager API
  version: 0.1.0
  description: REST API for managing IPsec VPN tunnels via strongSwan
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Tunnel:
      type: object
      required:
        - id
        - tunnel_id
        - name
        - peer_ip
        - local_subnet
        - auth_type
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          example: "1"
        tunnel_id:
          type: string
          example: "tun-abc123"
        name:
          type: string
          example: "my-vpn-tunnel"
        peer_ip:
          type: string
          example: "10.10.10.2"
        local_subnet:
          type: string
          example: "10.10.10.0/24"
        auth_type:
          type: string
          enum: [eap, l2tp, psk]
          example: "eap"
        psk:
          type: string
          description: Pre-shared key (only for PSK auth type)
          example: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4"
        username:
          type: string
          description: EAP/L2TP username
          example: "my-vpn-tunnel-aabb1122"
        status:
          type: string
          example: "active"
        metadata:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional user-provided metadata
          example: null
        created_at:
          type: string
          format: date-time
          example: "2026-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-15T10:30:00Z"
    CreateTunnelInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Tunnel name (1-128 chars, alphanumeric/spaces/hyphens/underscores)
          example: "my-vpn-tunnel"
        local_subnet:
          type: string
          description: "VPN subnet (default: 10.10.10.0/24)"
          example: "10.10.10.0/24"
        auth_type:
          type: string
          description: "Authentication type: eap (default), l2tp, or psk"
          enum: [eap, l2tp, psk]
          example: "eap"
        metadata:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional user-provided metadata
          example: null
    CreateTunnelResponse:
      type: object
      description: Tunnel data plus plaintext password (only for EAP/L2TP on create)
      allOf:
        - $ref: "#/components/schemas/Tunnel"
        - type: object
          properties:
            password:
              type: string
              description: Plaintext password (only returned on create for EAP/L2TP)
              example: "secret123"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "tunnel tun-abc123 not found"
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          example: "healthy"
        error:
          type: string
          description: Error details (only present when unhealthy)
          example: "connection refused"
    DataMessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
              example: "deleted"
security: [BearerAuth: []]

paths:
  /healthz:
    get:
      summary: Health check
      description: Returns the health status of the service including database connectivity.
      operationId: healthz
      tags: [Health]
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: healthy
        "503":
          description: Service is unhealthy (e.g., database unreachable)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: unhealthy
                error: "connection refused"

  /api/v1/tunnels:
    post:
      summary: Create a tunnel
      description: Creates a new VPN tunnel with an allocated IP, strongSwan config, and credentials.
      operationId: createTunnel
      tags: [Tunnels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTunnelInput"
      responses:
        "201":
          description: Tunnel created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CreateTunnelResponse"
        "400":
          description: Bad request (invalid JSON, missing name, or invalid name)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: No available IP addresses in pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: List tunnels
      description: Returns all tunnels ordered by creation time (newest first). Optionally filter by username.
      operationId: listTunnels
      tags: [Tunnels]
      parameters:
        - name: username
          in: query
          required: false
          description: Filter tunnels by username
          schema:
            type: string
      responses:
        "200":
          description: List of tunnels
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tunnel"
              example:
                data:
                  - id: "1"
                    tunnel_id: "tun-abc123"
                    name: "my-vpn-tunnel"
                    peer_ip: "10.10.10.2"
                    local_subnet: "10.10.10.0/24"
                    auth_type: "eap"
                    username: "my-vpn-tunnel-aabb1122"
                    status: "active"
                    created_at: "2026-01-15T10:30:00Z"
                    updated_at: "2026-01-15T10:30:00Z"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/tunnels/{tunnelID}:
    get:
      summary: Get a tunnel
      description: Returns a single tunnel by its tunnel ID.
      operationId: getTunnel
      tags: [Tunnels]
      parameters:
        - name: tunnelID
          in: path
          required: true
          description: Tunnel ID (e.g., tun-abc123)
          schema:
            type: string
      responses:
        "200":
          description: Tunnel data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Tunnel"
        "404":
          description: Tunnel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a tunnel
      description: Deletes a tunnel, its strongSwan configuration, IP allocation, and secrets.
      operationId: deleteTunnel
      tags: [Tunnels]
      parameters:
        - name: tunnelID
          in: path
          required: true
          description: Tunnel ID (e.g., tun-abc123)
          schema:
            type: string
      responses:
        "200":
          description: Tunnel deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataMessageResponse"
              example:
                data:
                  message: deleted
        "404":
          description: Tunnel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/tunnels/{tunnelID}/rsc:
    get:
      summary: Get MikroTik RSC script
      description: Returns a rendered MikroTik RouterOS import script for the tunnel.
      operationId: getMikroTikRSC
      tags: [Tunnels]
      parameters:
        - name: tunnelID
          in: path
          required: true
          description: Tunnel ID (e.g., tun-abc123)
          schema:
            type: string
      responses:
        "200":
          description: RouterOS script
          content:
            text/plain:
              schema:
                type: string
              example: |
                /ip address add address=10.10.10.2/24 interface=ether1
                /ip ipsec peer add name=tun-abc123
        "404":
          description: Tunnel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/reload:
    post:
      summary: Reload strongSwan
      description: Reloads the strongSwan daemon configuration.
      operationId: reloadAll
      tags: [System]
      responses:
        "200":
          description: Reloaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataMessageResponse"
              example:
                data:
                  message: reloaded
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"